BigDecimal fix 18:57

+++ fund-base/src/test/java/pl/js/fund/model/FundPriceProviderTest.java
@@ -1,5 +1,7 @@
 package pl.js.fund.model;
 
+import java.math.BigDecimal;
+
 import junit.framework.Assert;
 
 import org.joda.time.LocalDate;
@@ -69,7 +71,7 @@ public class FundPriceProviderTest
     {
         provider.setFund(fund);
 
-        Double price = provider.getPriceAtLastBusinessDay(new LocalDate(2014, 04, 14));
+        BigDecimal price = provider.getPriceAtLastBusinessDay(new LocalDate(2014, 04, 14));
 
         Assert.assertNotNull(price);
     }
